refactor(tauri): split runtime helpers from lib.rs
