+++ desktop/src-tauri/src/runtime/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod diagnostics;
+pub(crate) mod profile;
+pub(crate) mod provider;
+pub(crate) mod proxy;
+pub(crate) mod science;
+pub(crate) mod transaction;
