+++ desktop/src-tauri/src/runtime/science.rs
@@ -1,0 +1,92 @@
+use std::path::PathBuf;
+
+use crate::config;
+
+/// 沙箱可写工作目录（独立 HOME）：`~/.csswitch/sandbox/home`。
+pub(crate) fn sandbox_home() -> PathBuf {
+    config::default_dir().join("sandbox").join("home")
+}
+
+/// 端口变更是否需要拆掉现有链路（纯函数，P1-c）。代理/沙箱任一端口变了，正在跑的代理就绑在
+/// 旧端口、正在跑的沙箱又把旧代理 URL 烘死了，二者与新配置不一致 → 拆掉逼下次「一键开始」按新端口重建。
+pub(crate) fn settings_change_needs_teardown(
+    old_proxy: u16,
+    new_proxy: u16,
+    old_sandbox: u16,
+    new_sandbox: u16,
+) -> bool {
+    old_proxy != new_proxy || old_sandbox != new_sandbox
+}
+
+/// 从 `claude-science url` 的 stdout 里取**第一条**合法 http(s) URL。
+pub(crate) fn first_http_url(stdout: &str) -> Option<String> {
+    for line in stdout.lines() {
+        let t = line.trim();
+        if t.starts_with("http://") || t.starts_with("https://") {
+            let url = t.split_whitespace().next().unwrap_or(t);
+            return Some(url.to_string());
+        }
+    }
+    None
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{first_http_url, sandbox_home, settings_change_needs_teardown};
+
+    // ---------- P1-c: 端口变更是否需拆链路（纯函数，4 组合） ----------
+    #[test]
+    fn settings_teardown_when_any_port_changes() {
+        assert!(
+            !settings_change_needs_teardown(18991, 18991, 8990, 8990),
+            "端口未变 → 不拆链路"
+        );
+        assert!(
+            settings_change_needs_teardown(18991, 19000, 8990, 8990),
+            "代理端口变 → 拆（旧代理绑旧端口、沙箱烘旧 URL）"
+        );
+        assert!(
+            settings_change_needs_teardown(18991, 18991, 8990, 9000),
+            "沙箱端口变 → 拆（旧沙箱在旧端口成孤儿）"
+        );
+        assert!(
+            settings_change_needs_teardown(18991, 19000, 8990, 9000),
+            "都变 → 拆"
+        );
+    }
+
+    #[test]
+    fn first_http_url_takes_only_first_valid_url() {
+        let multi = "http://127.0.0.1:8990/setup?nonce=abc123\n\
+                     This is a single-use link, expires in 60 seconds.";
+        assert_eq!(
+            first_http_url(multi).as_deref(),
+            Some("http://127.0.0.1:8990/setup?nonce=abc123"),
+        );
+        let inline = "https://x.example/y?z=1  (single-use)";
+        assert_eq!(
+            first_http_url(inline).as_deref(),
+            Some("https://x.example/y?z=1")
+        );
+        let lead = "Open this link in your browser:\nhttp://127.0.0.1:8990/a";
+        assert_eq!(
+            first_http_url(lead).as_deref(),
+            Some("http://127.0.0.1:8990/a")
+        );
+        assert_eq!(first_http_url("no url here\nnor here"), None);
+        assert_eq!(
+            first_http_url("http://127.0.0.1:8990").as_deref(),
+            Some("http://127.0.0.1:8990")
+        );
+    }
+
+    #[test]
+    fn sandbox_home_is_writable_under_config_dir() {
+        let h = sandbox_home();
+        assert!(h.ends_with("sandbox/home"), "应以 sandbox/home 结尾：{h:?}");
+        assert!(
+            h.to_string_lossy().contains(".csswitch"),
+            "应在 .csswitch 下：{h:?}"
+        );
+    }
+}
